Pull DescriptorInfo to top level and generify to MethodDescriptor.
